Create Planning Material
This is my deliverable, as I used this as a self guide on where I wanted to take the next step in the project. I have used VSC on the whole project and didn't transfer over until near completion. I see now what impact on how helpful utilizing git the whole project takes.